add test-provider script and on-demand GitHub workflow
Add test-provider.ts script that accepts a --target scrape target string and
validates provider output (text extraction, citations, rawOutput round-trip).

Add test-providers.yaml workflow_dispatch workflow that runs all main targets
as parallel matrix jobs, or a single target when specified.

Remove resolveProviderId (was a pointless pass-through).
